BASE: Move EngineMan code to dedicated files
