+++ base/commandLine.cpp
@@ -27,6 +27,7 @@
 
 #include <limits.h>
 
+#include "engines/engineman.h"
 #include "engines/metaengine.h"
 #include "base/commandLine.h"
 #include "base/plugins.h"
