+++ engines/engine.cpp
@@ -27,6 +27,7 @@
 #endif
 
 #include "engines/engine.h"
+#include "engines/engineman.h"
 #include "engines/dialogs.h"
 #include "engines/util.h"
 #include "engines/metaengine.h"
