+++ engines/metaengine.h
@@ -563,61 +563,4 @@ public:
 	static WARN_UNUSED_RESULT bool readSavegameHeader(Common::InSaveFile *in, ExtendedSavegameHeader *header, bool skipThumbnail = true);
 };
 
-/**
- * Singleton class that manages all engine plugins.
- */
-class EngineManager : public Common::Singleton<EngineManager> {
-public:
-	/**
-	 * Given a list of FSNodes in a given directory, detect a set of games contained within.
-	 *
-	 * Returns an empty list if none are found.
-	 */
-	DetectionResults detectGames(const Common::FSList &fslist) const;
-
-	/** Find a plugin by its engine ID. */
-	const Plugin *findPlugin(const Common::String &engineId) const;
-
-	/**
-	 * Get the list of all plugins for the type specified.
-	 *
-	 * By default, it will get METAENGINES, for now.
-	 * If usage of actual engines never occurs, the default arguments can be skipped,
-	 * and always have it return PLUGIN_TYPE_ENGINE_DETECTION.
-	 */
-	const PluginList &getPlugins(const PluginType fetchPluginType = PLUGIN_TYPE_ENGINE_DETECTION) const;
-
-	/** Find a target. */
-	QualifiedGameDescriptor findTarget(const Common::String &target, const Plugin **plugin = NULL) const;
-
-	/**
-	 * List games matching the specified criteria.
-	 *
-	 * If the engine ID is not specified, this scans all the plugins,
-	 * loading them from the disk if necessary. This is a slow operation on
-	 * some platforms and should not be used for the happy path.
-	 */
-	QualifiedGameList findGamesMatching(const Common::String &engineId, const Common::String &gameId) const;
-
-	/**
-	 * Create a target from the supplied game descriptor.
-	 *
-	 * @return The created target name.
-	 */
-	Common::String createTargetForGame(const DetectedGame &game);
-
-	/** Upgrade a target to the current configuration format. */
-	void upgradeTargetIfNecessary(const Common::String &target) const;
-
-private:
-	/** Find a game across all loaded plugins. */
-	QualifiedGameList findGameInLoadedPlugins(const Common::String &gameId) const;
-
-	/** Use heuristics to complete a target lacking an engine ID. */
-	void upgradeTargetForEngineId(const Common::String &target) const;
-};
-
-/** Convenience shortcut for accessing the engine manager. */
-#define EngineMan EngineManager::instance()
-/** @} */
 #endif
